+++ src/main/java/com/ndb/auction/dao/oracle/transactions/stripe/StripeDepositDao.java
@@ -20,7 +20,7 @@ public class StripeDepositDao extends BaseOracleDao implements IStripeDao {
         StripeDepositTransaction m = new StripeDepositTransaction();
         m.setId(rs.getInt("ID"));
         m.setUserId(rs.getInt("USER_ID"));
-        m.setAmount(rs.getLong("AMOUNT"));
+        m.setAmount(rs.getDouble("AMOUNT"));
         m.setCreatedAt(rs.getTimestamp("CREATED_AT").getTime());
         m.setConfirmedAt(rs.getTimestamp("UPDATED_AT").getTime());
         m.setStatus(rs.getBoolean("STATUS"));
