+++ src/main/java/com/ndb/auction/dao/oracle/transactions/stripe/StripePresaleDao.java
@@ -33,7 +33,7 @@ public class StripePresaleDao extends BaseOracleDao implements ITransactionDao {
 		m.setUserId(rs.getInt("USER_ID"));
 		m.setOrderId(rs.getInt("ORDER_ID"));
         m.setPresaleId(rs.getInt("PRESALE_ID"));
-		m.setAmount(rs.getLong("AMOUNT"));
+		m.setAmount(rs.getDouble("AMOUNT"));
 		m.setCreatedAt(rs.getTimestamp("CREATED_AT").getTime());
         m.setConfirmedAt(rs.getTimestamp("UPDATED_AT").getTime());
 		m.setStatus(rs.getBoolean("STATUS"));
